requirements.txt
absl-py==1.4.0
Adafruit-Blinka==8.16.1
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-requests==1.13.0
adafruit-circuitpython-typing==1.8.3
Adafruit-PlatformDetect==3.42.0
Adafruit-PureIO==1.1.10
arandr==0.1.10
astroid==2.5.1
asttokens==2.0.4
astunparse==1.6.3
automationhat==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
board==1.0
buttonshim==0.0.2
cachetools==5.3.0
Cap1xxx==0.1.3
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
deepface==0.0.79
distro==1.5.0
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
filelock==3.10.0
fire==0.5.0
Flask==1.1.2
flatbuffers==20181003210633
fourletterphat==0.1.0
gast==0.4.0
gdown==4.6.4
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpiozero==1.6.2
grpcio==1.51.3
gunicorn==20.1.0
h5py==3.8.0
html5lib==1.1
idna==2.10
importlib-metadata==6.0.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
keras==2.11.0
lazy-object-proxy==0.0.0
libclang==15.0.6.1
logilab-common==1.8.1
lxml==4.6.3
Markdown==3.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mtcnn==0.1.1
mypy==0.812
mypy-extensions==0.4.3
numpy==1.24.2
oauthlib==3.1.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pantilthat==0.0.7
parso==0.8.1
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera2==0.3.9
pidng==4.0.9
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
protobuf==3.19.6
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycups==2.0.1
pyftdi==0.54.0
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
pyngrok==5.2.1
PyOpenGL==3.1.5
pyOpenSSL==20.0.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5b0
pysmbc==1.0.23
PySocks==1.7.1
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
pytz==2022.7.1
pyusb==1.2.1
PyYAML==6.0
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
retina-face==0.0.13
roman==2.0.0
rpi-ws281x==4.3.4
RPi.GPIO==0.7.0
rsa==4.9
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
Send2Trash==1.6.0b1
sense-hat==2.4.0
simplejpeg==1.6.4
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-cpu-aws==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
thonny==4.0.1
toml==0.10.1
touchphat==0.0.1
tqdm==4.65.0
twython==3.8.2
typed-ast==1.4.2
typing-extensions==4.5.0
unicornhathd==0.0.4
urllib3==1.26.5
v4l2-python3==0.3.2
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.15.0
